pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdhistogram"
version = "1.0.0"
description = "Google Docs Revision History Analyzer - Analyze typing behavior and detect anomalies"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "GDHistogram Team"}
]
keywords = [
    "google-docs",
    "revision-history",
    "typing-analysis",
    "wpm",
    "anomaly-detection",
    "histogram"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "PySide6>=6.6.0",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
    "cryptography>=41.0.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "mypy>=1.6.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/NagusameCS/GDHistogram"
Repository = "https://github.com/NagusameCS/GDHistogram.git"
Issues = "https://github.com/NagusameCS/GDHistogram/issues"

[project.scripts]
gdhistogram = "gdhistogram.main:main"

[project.gui-scripts]
gdhistogram-gui = "gdhistogram.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gdhistogram*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
